refactor(#234) : 파일구조 변경
